+++ CapstoneProject/Controllers/CoatsController.cs
@@ -54,15 +54,31 @@ namespace CapstoneProject.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,TypeId,Brand,Description,Size")] Coat coat)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (ModelState.IsValid)
+            if (id == null)
             {
-                _context.Add(coat);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return NotFound();
             }
-            return View(coat);
+            var coatToUpdate = await _context.Coats.FirstOrDefaultAsync(s => s.Id == id);
+            if (await TryUpdateModelAsync<Coat>(
+                coatToUpdate,
+                "",
+                s => s.Brand, s => s.Description, s => s.Size))
+            {
+                try
+                {
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
+                }
+            }
+            return View(coatToUpdate);
         }
 
         // GET: Coats/Edit/5
@@ -117,7 +133,7 @@ namespace CapstoneProject.Controllers
         }
 
         // GET: Coats/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -125,12 +141,19 @@ namespace CapstoneProject.Controllers
             }
 
             var coat = await _context.Coats
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (coat == null)
             {
                 return NotFound();
             }
 
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
             return View(coat);
         }
 
@@ -140,14 +163,20 @@ namespace CapstoneProject.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var coat = await _context.Coats.FindAsync(id);
-            _context.Coats.Remove(coat);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool CoatExists(int id)
-        {
-            return _context.Coats.Any(e => e.Id == id);
+            if (coat == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Coats.Remove(coat);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
     }
 }
