+++ CapstoneProject/Controllers/SnowboardsController.cs
@@ -54,13 +54,20 @@ namespace CapstoneProject.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,TypeId,Brand,Description,Size")] Snowboard snowboard)
+        public async Task<IActionResult> Create([Bind("TypeId,Brand,Description,Size")] Snowboard snowboard)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(snowboard);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+            if (ModelState.IsValid)
+                {
+                    _context.Add(snowboard);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException)
+            {
+                ModelState.AddModelError("", "Unable to save changes. " + "see your system administrator.");
             }
             return View(snowboard);
         }
@@ -86,38 +93,35 @@ namespace CapstoneProject.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,TypeId,Brand,Description,Size")] Snowboard snowboard)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != snowboard.Id)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var snowboardToUpdate = await _context.Snowboards.FirstOrDefaultAsync(s => s.Id == id);
+            if (await TryUpdateModelAsync<Snowboard>(
+                snowboardToUpdate,
+                "",
+                s => s.Brand, s => s.Description, s => s.Size))
             {
                 try
                 {
-                    _context.Update(snowboard);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!SnowboardExists(snowboard.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(snowboard);
+            return View(snowboardToUpdate);
         }
 
         // GET: Snowboards/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -125,12 +129,18 @@ namespace CapstoneProject.Controllers
             }
 
             var snowboard = await _context.Snowboards
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (snowboard == null)
             {
                 return NotFound();
             }
-
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
             return View(snowboard);
         }
 
@@ -140,14 +150,20 @@ namespace CapstoneProject.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var snowboard = await _context.Snowboards.FindAsync(id);
-            _context.Snowboards.Remove(snowboard);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool SnowboardExists(int id)
-        {
-            return _context.Snowboards.Any(e => e.Id == id);
+            if (snowboard == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Snowboards.Remove(snowboard);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
     }
 }
